requirements.txt
discord.io==0.8.1